pyproject.toml
[tool.poetry]
name = "datacity-k8s"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ruamel-yaml = "^0.18.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
